Manage exceptions in service layer.

+++ src/main/java/ma/formations/rest/service/ServiceImpl.java
@@ -17,7 +17,11 @@ public class ServiceImpl implements IService {
 
     @Override
     public ArticleDTO getById(Long id) {
-        return ArticleConverter.toDTO(dao.findById(id));
+        Article articleFound = dao.findById(id);
+        if (articleFound == null) {
+            throw new BusinessException(String.format("no article with id= %s exist", id));
+        }
+        return ArticleConverter.toDTO(articleFound);
     }
 
     @Override
@@ -27,20 +31,40 @@ public class ServiceImpl implements IService {
 
     @Override
     public void create(ArticleDTO article) {
-        Article articleFound = dao.findAll().stream().filter(a -> article.getId().equals(a.getId())).findFirst().orElse(null);
+        if (article.getId() == null) {
+            dao.save(ArticleConverter.toBO(article));
+            return;
+        }
+        Article articleFound = dao.findAll().stream().
+                filter(a -> article.getId().equals(a.getId())).
+                findFirst().orElse(null);
+
         if (articleFound != null)
-            throw new BusinessException("Article with the same Id=" + article.getId() + " exist in database");
+            throw new BusinessException(String.format("Article with the same Id=%s exist in database", article.getId()));
+
         dao.save(ArticleConverter.toBO(article));
     }
 
     @Override
     public void update(Long id, ArticleDTO article) {
+        if (id == null)
+            throw new BusinessException(String.format("Article id=%s should not be null", id));
+
+        dao.findAll().stream().
+                filter(a -> a.getId().equals(id)).
+                findFirst().orElseThrow(() -> new BusinessException(String.format("No article with the id=%s exist in database", id)));
+
         article.setId(id);
         dao.save(ArticleConverter.toBO(article));
+        
     }
 
     @Override
     public void deleteById(Long id) {
+        Article articleFound = dao.findById(id);
+        if (articleFound == null) {
+            throw new BusinessException(String.format("no article with id= %s exist", id));
+        }
         dao.deleteById(id);
     }
 }
